@@ -1,8 +1,9 @@
+"use client";
+
 import React, { useContext } from "react";
-import type { NextPage } from "next";
 import styles from "../styles/Home.module.css";
 
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import HowToParticipate from "../components/pages/home/howToParticipate";
 import QuestCategories from "../components/pages/home/questCategories";
 import TrendingQuests from "../components/pages/home/trending";
@@ -10,7 +11,7 @@ import Blur from "../components/shapes/blur";
 import { QuestsContext } from "../context/QuestsProvider";
 import FeaturedQuest from "../components/UI/featured_banner/featuredQuest";
 
-const Quests: NextPage = () => {
+export default function Page() {
   const router = useRouter();
   const { featuredQuest, categories, trendingQuests } =
     useContext(QuestsContext);
@@ -47,6 +48,4 @@ const Quests: NextPage = () => {
       </div>
     </div>
   );
-};
-
-export default Quests;
+}
