@@ -8,16 +8,17 @@ type QuestCategoryProps = {
 
 const QuestCategory: FunctionComponent<QuestCategoryProps> = ({ category }) => {
   return (
-    <Link href={`/categories/${category.name}`}>
-      <div className={styles.questCategory}>
-        <div className={styles.categoryInfos}>
-          <h2 className="text-gray-200">{category.name} Quest</h2>
-          <p className="text-gray-200">
-            {category.questNumber} quest{category.questNumber > 1 ? "s" : null}
-          </p>
-        </div>
-        <img src={category.img} />
+    <Link
+      href={`/categories/${category.name}`}
+      className={styles.questCategory}
+    >
+      <div className={styles.categoryInfos}>
+        <h2 className="text-gray-200">{category.name} Quest</h2>
+        <p className="text-gray-200">
+          {category.questNumber} quest{category.questNumber > 1 ? "s" : null}
+        </p>
       </div>
+      <img src={category.img} />
     </Link>
   );
 };
