@@ -1,7 +1,11 @@
-import React from "next";
+import React, {
+  FunctionComponent,
+  ReactNode,
+  useEffect,
+  useState,
+} from "react";
 import StartScreen from "./startScreen";
 import styles from "../../styles/components/quests/quiz.module.css";
-import { FunctionComponent, ReactNode, useEffect, useState } from "react";
 import Step from "./step";
 import EndScreen from "./endScreen";
 import QuizLoading from "../../components/quiz/quizLoading";
